Example DataVolley files provided as part of the datavolley package

Example DataVolley files provided as part of the datavolley package

dv_example_file(choice = 1)

Arguments

choice

numeric: which data file to return?

1 - the January 2015 Slovenian junior women's final between Braslovče and Nova KBM Branik

2 - the December 2012 men's Slovenian national championship semifinal between ACH Volley and Maribor

Value

path to the file

References

The example data files came from http://www.odbojka.si/

See also

Examples

myfile <- dv_example_file()
x <- dv_read(myfile, insert_technical_timeouts = FALSE)
summary(x)
#> Match summary:
#> Date: 2015-01-25
#> League: Finale mladinke
#> Teams: Braslovče (JERONČIČ ZORAN/MIHALINEC DAMIJANA)
#>        vs
#>        Nova KBM Branik (HAFNER MATJAŽ)
#> Result: 3-0 (25-16, 25-14, 25-22)
#> Duration: 67 minutes