Summarize the video sync times in a dvw file

This function will generate a summary of various video time differences in a dvw file. Apply this to a file that you have synchronized to video, and the results can be used to tweak the behaviour of dv_sync_video.

Usage

dv_sync_summary(x)

Arguments

x

datavolley: a single datavolley object as returned by dv_read, or the plays component of one

Value

A data.frame with columns type, N, mean, most_common, min, max

See also

Examples

x <- dv_read(dv_example_file(3))
dv_sync_summary(x)
#> # A tibble: 12 × 6
#>    type                                      N  mean most_common   min   max
#>    <chr>                                 <int> <dbl>       <dbl> <dbl> <dbl>
#>  1 Serve to receive time                    97 2.04            2    -1     6
#>  2 Serve to set time (Jump serve)           20 4.9             5     3    10
#>  3 Serve to set time (Jump-float serve)     36 3.83            3     1     7
#>  4 Serve to set time (Other serve types)    16 3.19            3     2     8
#>  5 Set to attack time (High ball set)       54 0.889           0    -1     3
#>  6 Set to attack time (Other set types)     57 1.04            1    -1     3
#>  7 Set to attack time (Quick ball set)      16 0.812           2    -5     3
#>  8 Attack to block time                     40 1.8             3     0     7
#>  9 Dig to set time                          39 0.513           1    -2     2
#> 10 Attack to counter-attack time            48 3.69            3     0     6
#> 11 Overpass to next touch time              13 4.15            5     0     8
#> 12 Last touch to end of rally              107 0               0     0     0