Synchronize video times

Warning: experimental! This function uses the scouted clock time of each serve and some simple rules to align each scouted contact with its (approximately correct) time in the corresponding match video.

Usage

dv_sync_video(
  x,
  first_serve_time,
  freeball_digs_accurate = FALSE,
  config = dv_sync_video_config()
)

dv_sync_video_config()

Arguments

x

datavolley: a single datavolley object as returned by dv_read

first_serve_time

numeric or string: the time of the first serve in the video. This can be a numeric value giving the time in seconds, or a string of the form "MM:SS" (minutes and seconds) or "HH:MM:SS" (hours, minutes and seconds)

freeball_digs_accurate

logical: if TRUE, treat the clock time of each freeball dig as being accurate (in the same way that serve times are treated as being accurate). Obviously this requires that the scout has been accurate in their timing when entering freeball digs, but assuming that is the case then setting freeball_digs_accurate to TRUE will improve the synchronization of rallies with freeballs (which otherwise tends to be poor, because the play is messy and less predictable compared to in-system rallies)

config

list: a set of parameters that control the syncing process. See Details, below

Value

A modified copy of x

Details

When a match is scouted live, the clock time of each serve will usually be correct because the scout can enter the serve code at the actual time of serve. But the remainder of the touches in the rally might not be at the correct relative times, if the scout can't keep up with the live action. This function uses some simple rules to better synchronize the scouted contacts and corresponding match video.

The config object contains a set of times (in seconds), which you can adjust to suit your scouting style and level of play. If you have an already-synchronized dvw file, the dv_sync_summary function can provide some guidance as to what these values should be. The config object contains the following entries:

S_R - the time between serve and reception

S_E - the time between serve and opposition first set (but see also SQ and SM, below)

E_A - the time between set and attack (but see also EQ and EH, below)

A_B - the time between attack and block

A_A - the time between attacks in the rally (i.e. attack to counter-attack)

D_E - the time between dig and set

RDov - the time between reception or dig overpass and the next touch by the opposition

END - the time between the last contact and end-of-rally marker

SQ - S_E adjustment for jump serves (this value gets added to S_E for jump serves)

SM - S_E adjustment for jump-float serves (this value gets added S_E for jump-float serves)

EQ - E_A adjustment for quick sets (this value gets added to E_A for quick sets)

EH - E_A adjustment for high sets (this value gets added to E_A for high sets)

See also

Examples

x <- dv_read(dv_example_file())
## first serve was at 54s in the video
x <- dv_sync_video(x, first_serve_time = 54)

## with a custom configuration
my_config <- dv_sync_video_config()
my_config$S_A <- 4 ## change entries as necessary
## first serve was at 3:35 in the video
x <- dv_sync_video(x, first_serve_time = "3:35", config = my_config)