Find a particular match in a list of datavolley objects

Find a particular match in a list of datavolley objects

Usage

find_match(match_id, x)

Arguments

match_id

string: match_id to find

x

list: list of datavolley objects as returned by dv_read

Value

numeric index of the match in the list

See also