Change team names

A team name can sometimes be spelled incorrectly, particularly if there are character encoding issues. This can be a particular problem when combining data from multiple files. A team name matching the from entry in a row in remap is renamed to the corresponding to value. If the remap object also contains a team_id column, then a team name must match the team_id and from values to be changed to the corresponding to value.

remap_team_names(x, remap, fixed = TRUE)

Arguments

x

datavolley: a datavolley object as returned by dv_read, or list of such objects

remap

data.frame: data.frame of strings with columns from and to, and optionally team_id

fixed

logical: treat the to and from entries as fixed string? (Treat as regexps if fixed is FALSE)

Value

datavolley object or list with corresponding team names changed

See also

Examples

if (FALSE) {
  x <- dv_read(dv_example_file(), insert_technical_timeouts=FALSE)
  summary(x)

  ## rename a team based just on team name
  summary(remap_team_names(x, data.frame(from="Nova KBM Branik", to="NKBM Branik")))

  ## rename a team based on team name and ID
  summary(remap_team_names(x, data.frame(from="Nova KBM Branik", to="NKBM Branik", id="MB4")))
}